remove: 기존 디렉토리 제거, MVC구조로 대체
