+++ src/view/GameView.js
